разделил на модули
так же изменил метод создания датафреймов через библиотеку matplotlib а не через модуль csv. изменил функцию install_graphs на установку единичного графика. убрал функцию установки 3 датафреймов. теперь нужно вводить 3 пути для установки графиков на основе датафреймов.
